rename type_promotion_table.h -> data_type_promotion.h
